+++ client/tws_foundation_client/lib/src/services/business/bases/drivers_service_base.dart
@@ -10,4 +10,10 @@ abstract class DriversServiceBase extends TWSServiceBase {
 
   /// Transaction to generate a set view object.
   Effect<SetViewOut<Driver>> view(SetViewOptions<Driver> options, String auth);
+
+  /// Transaction to create a set view object.
+  Effect<SetBatchOut<Driver>> create(List<Driver> drivers, String auth);
+
+  /// Transaction to update a set object.
+  Effect<RecordUpdateOut<Driver>> update(Driver driver, String auth);
 }