+++ client/tws_foundation_client/lib/src/sets/business/driver_external.dart
@@ -2,29 +2,57 @@ import 'package:csm_client/csm_client.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 final class DriverExternal implements CSMSetInterface {
+
+  /// [Identification] property key.
   static const String kIdentification = "identification";
+
+  /// [common] property key.
   static const String kCommon = "common";
+
+  /// [DriverCommon] property key.
   static const String kDriverCommonNavigation = 'DriverCommonNavigation';
+
+  /// [identificationNavigation] property key.
   static const String kIdentificationNavigation = 'IdentificationNavigation';
 
   late final DateTime _timestamp;
   DateTime get timestamp => _timestamp;
-  
+
+  /// Record database pointer.
   @override
   int id = 0;
+
+  /// Foreign  relation [Status] key.
   int status = 1;
+
+  /// Foreign relation [Identification] key.
   int identification = 0;
+
+  /// Foreign relation [DriverCommon] key.
   int common = 0;
+
+  /// [DriverCommon] Navigation set. 
   DriverCommon? driverCommonNavigation;
+
+  /// [Identification] Navigation set.
   Identification? identificationNavigation;
+
+  /// [Status] Navigation set.
   Status? statusNavigation;
 
+  /// Creates an [DriverExternal] object with default values.
+  DriverExternal.a(){
+    _timestamp = DateTime.now();
+  }
+
+  /// Creates an [DriverExternal] object with requiered properties.
   DriverExternal(this.id, this.status, this.identification, this.common, this.driverCommonNavigation, this.identificationNavigation, this.statusNavigation, { 
     DateTime? timestamp,
   }){
     _timestamp = timestamp ?? DateTime.now(); 
   }
 
+  /// Creates an [DriverExternal] object based on a serialized json object.
   factory DriverExternal.des(JObject json) {
     int id = json.get(SCK.kId);
     int status = json.get(SCK.kStatus);
@@ -79,7 +107,8 @@ final class DriverExternal implements CSMSetInterface {
 
     return results;
   }
-  DriverExternal.def();
+
+  /// Creates a clone for a [DriverExternal] object, overriding the given values.
   DriverExternal clone({
     int? id,
     int? status,
@@ -89,6 +118,12 @@ final class DriverExternal implements CSMSetInterface {
     DriverCommon? driverCommonNavigation,
     Status? statusNavigation,
   }){
+
+    if (identification == 0) {
+      this.identificationNavigation = null;
+      identificationNavigation = null;
+    }
+ 
     return DriverExternal(id ?? this.id, status ?? this.status, identification ?? this.identification, common ?? this.common, driverCommonNavigation ?? this.driverCommonNavigation, identificationNavigation ?? this.identificationNavigation, statusNavigation ?? this.statusNavigation);
   }
 }