+++ client/tws_foundation_client/lib/src/tws_foundation_source.dart
@@ -1,21 +1,20 @@
 import 'package:csm_client/csm_client.dart';
-import 'package:tws_foundation_client/src/services/business/bases/trailers_classes_service_base.dart';
-import 'package:tws_foundation_client/src/services/business/bases/trailers_types_service_base.dart';
 import 'package:tws_foundation_client/src/services/business/carriers_service.dart';
 import 'package:tws_foundation_client/src/services/business/drivers_externals_service.dart';
 import 'package:tws_foundation_client/src/services/business/drivers_service.dart';
+import 'package:tws_foundation_client/src/services/business/employee_service.dart';
 import 'package:tws_foundation_client/src/services/business/load_type_service.dart';
 import 'package:tws_foundation_client/src/services/business/manufacturers_service.dart';
 import 'package:tws_foundation_client/src/services/business/sections_service.dart';
 import 'package:tws_foundation_client/src/services/business/situations_service.dart';
 import 'package:tws_foundation_client/src/services/business/trailer_service.dart';
+import 'package:tws_foundation_client/src/services/business/trailers_classes_service.dart';
 import 'package:tws_foundation_client/src/services/business/trailers_externals_service.dart';
+import 'package:tws_foundation_client/src/services/business/trailers_types_service.dart';
 import 'package:tws_foundation_client/src/services/business/trucks_externals_service.dart';
 import 'package:tws_foundation_client/src/services/business/trucks_inventories_service.dart';
 import 'package:tws_foundation_client/src/services/business/trucks_service.dart';
-import 'package:tws_foundation_client/src/services/business/trailers_classes_service.dart';
 import 'package:tws_foundation_client/src/services/business/vehicules_models_service.dart';
-import 'package:tws_foundation_client/src/services/business/trailers_types_service.dart';
 import 'package:tws_foundation_client/src/services/business/yard_log_service.dart';
 import 'package:tws_foundation_client/src/services/security/account_service.dart';
 import 'package:tws_foundation_client/src/services/security/contacts_service.dart';
@@ -78,13 +77,18 @@ final class TWSFoundationSource extends CSMSourceBase {
 
   /// Truck Inventory service
   late final TrucksInventoriesServiceBase trucksInventories;
+
   // Trailer Type models service
   late final TrailersTypesServiceBase trailersTypes;
+  
   // Trailer Class service
   late final TrailersClassesServiceBase trailersClasses;
 
   /// Contact service
   late final ContactsServiceBase contacts;
+
+  /// Employees service
+  late final EmployeeService employees;
 
   /// [TWSFoundationSource] instance constructor.
   ///
@@ -122,6 +126,7 @@ final class TWSFoundationSource extends CSMSourceBase {
     TrailersTypesService? trailersTypes,
     TrailersClassesService? trailersClasses,
     ContactsService? contacts,
+    EmployeeService? employees,
     CSMUri development = const CSMUri(
       '127.0.0.1',
       '',
@@ -155,5 +160,6 @@ final class TWSFoundationSource extends CSMSourceBase {
     this.trailersTypes = trailersTypes ?? TrailersTypesService(host, client: client);
     this.trailersClasses = trailersClasses ?? TrailersClassesService(host, client: client);    
     this.contacts = contacts ?? ContactsService(host, client: client);
+    this.employees = employees ?? EmployeeService(host, client: client);
   }
 }