+++ client/tws_foundation_client/test/definition/drivers_externals_service_test.dart
@@ -7,21 +7,29 @@ import 'package:tws_foundation_client/tws_foundation_client.dart';
 void main() {
   late DriversExternalsServiceBase service;
   late SetViewOut<DriverExternal> viewMock;
-  late DriverExternal createMock;
+  late SetBatchOut<DriverExternal> createMock;
+  late RecordUpdateOut<DriverExternal> updateMock;  
   late SetViewOptions<DriverExternal> options;
+  late List<DriverExternal> drivers;
+
 
   setUp(
     () {
       List<SetViewOrderOptions> noOrderigns = <SetViewOrderOptions>[];
       options = SetViewOptions<DriverExternal>(false, 10, 1, null, noOrderigns, <SetViewFilterNodeInterface<DriverExternal>>[]);
       viewMock = SetViewOut<DriverExternal>(<DriverExternal>[], 1, DateTime.now(), 3, 0, 20);
-      createMock = DriverExternal(0, 1, 1, 1, null, null, null);
-
+      createMock = SetBatchOut<DriverExternal>(<DriverExternal>[], <SetOperationFailure<DriverExternal>>[], 0, 0, 0, false);
+      updateMock = RecordUpdateOut<DriverExternal>(DriverExternal.a(), DriverExternal.a());
+      drivers = <DriverExternal>[
+        DriverExternal.a(),
+      ];
       Client mockClient = MockClient(
         (Request request) async {
           JObject jObject = switch (request.url.pathSegments.last) {
             'view' => SuccessFrame<SetViewOut<DriverExternal>>('qTracer', viewMock).encode(),
-            'create' => SuccessFrame<DriverExternal>('qTracer', createMock).encode(),
+            'create' => SuccessFrame<SetBatchOut<DriverExternal>>('qTracer', createMock).encode(),
+             'update' => SuccessFrame<RecordUpdateOut<DriverExternal>>('qTracer', updateMock).encode(),
+
             _ => <String, dynamic>{},
           };
 
@@ -66,4 +74,49 @@ void main() {
     },
     timeout: Timeout.factor(5),
   );
+  
+  test(
+    'Create',
+    () async {
+      MainResolver<SetBatchOut<DriverExternal>> fact = await service.create(drivers, '');
+      bool pased = false;
+      fact.resolve(
+        decoder: (JObject json) => SetBatchOut<DriverExternal>.des(json, DriverExternal.des),
+        onConnectionFailure: () {},
+        onFailure: (FailureFrame failure, int status) {
+          throw failure;
+        },
+        onSuccess: (SuccessFrame<SetBatchOut<DriverExternal>> success) {
+          pased = true;
+        },
+        onException: (Object exception, StackTrace trace) {
+          throw exception;
+        },
+      );
+      expect(pased, true);
+    },
+  );
+
+  test(
+    'Update',
+    () async {
+      MainResolver<RecordUpdateOut<DriverExternal>> fact = await service.update(DriverExternal.a(), '');
+      bool pased = false;
+      fact.resolve(
+        decoder: (JObject json) => RecordUpdateOut<DriverExternal>.des(json, DriverExternal.des),
+        onConnectionFailure: () {},
+        onFailure: (FailureFrame failure, int status) {
+          throw failure;
+        },
+        onSuccess: (SuccessFrame<RecordUpdateOut<DriverExternal>> success) {
+          pased = true;
+        },
+        onException: (Object exception, StackTrace trace) {
+          throw exception;
+        },
+      );
+      expect(pased, true);
+    },
+  );
+
 }